Reorganise parser code and reduce visible scope
